updated hausdorff generator and hausdorff script code to keep a row/col of nan's corresponding to the participant removed, enabling easier comparison between hausdorff matrices. Note that in distance_metrics.py, the column and row labels had to be integers

+++ src/distance_metrics.py
@@ -12,7 +12,7 @@ from sklearn.neighbors import NearestNeighbors
 
 def hausdorff_distances(emb_2d):
     # directed or assymetric variant     
-    labels = emb_2d.part_id.unique()
+    labels = emb_2d.part_id.unique().astype('int32')
     len_labels = len(labels)
     # build empty df
     pairwise_distances_hausdorff = pd.DataFrame(np.zeros((len_labels, len_labels)) , columns = labels, index=labels)
